rename user fields
